inc.ClassDMS.php - Added a new method to retrieve the next memo number given a user and fixed bug where the getDocumetnIDByNumber method wasn't updated for the new name in a mysql table column.
inc.ClassFolder.php - Updated methodology to generate memo numbers. A new indexNumber column was added to tblMemoNumbers to contain the numeric portion of the memo number. The maximum value in the indexNumber column is found and 1 is added to it to automatically generate the next memo number. However, if a docNumber is passed in the addDocument method, it will be used instead only if it does not already exist.

op.AddDocument.php - Added a check to get post data if set for the document number. If set, pass it to the addDocument method.

class.AddDocument.php - Add an input to display the next memo number and allow user to modify it before submitting memo.

class.Search.php - Removed status column.